[adds] + Ability to load MCM data pre-injection or post-injection (default) + UIRoot_Entry.ShouldShowMainMenuGUI detour to skip calling MainMenuDrawer in additional situations
[changes]
+ Too many "StaticClasses", structured into Classes/ Classes/Abstract and Classes/Static
+ MainMenuDrawer detours simplified and supporting methods and properties moved to public API
+ Library sub-controller doesn't hard-fail if Core and CCL are out of order, handled by Window_WarnRestart instead
+ Removed unintentional "s" from MenuWorker"s"
+ Increased size of Window_WarnRestart from 400x300 -> 600x500

[fixes]
+ Language icon texture on main menu
