+++ DLL_Project/Classes/Abstract/SpecialInjector.cs
