+++ DLL_Project/Classes/HelpDetailSectionHelper.cs
