+++ DLL_Project/Classes/Static/CCL_Log.cs
