+++ DLL_Project/Classes/Static/Detours.cs
