+++ DLL_Project/Classes/Static/HelpBuilder.cs
