+++ DLL_Project/Classes/Static/InterModCommunication.cs
