+++ DLL_Project/Classes/Static/Research.cs
