+++ DLL_Project/Defs/ModHelperDef.cs
@@ -21,6 +21,8 @@ namespace CommunityCoreLibrary
 
         public Verbosity                    Verbosity = Verbosity.Default;
 
+        // TODO:  Add ability to load MCM data before anything else (default behaviour: last as per 0.13.1 and older)
+        public bool                         PreloadMCMs = false;
         public List< MCMInjectionSet >      ModConfigurationMenus;
 
         #endregion
